% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\docType{data}
\name{CLINICAL_ALIASES}
\alias{CLINICAL_ALIASES}
\title{Clinical variable aliases (smart matching)}
\format{
An object of class \code{list} of length 29.
}
\usage{
CLINICAL_ALIASES
}
\description{
Clinical variable aliases (smart matching)
}
\keyword{internal}
