% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\docType{data}
\name{CLINICAL_GROUPS}
\alias{CLINICAL_GROUPS}
\title{Clinical variable groups}
\format{
An object of class \code{list} of length 4.
}
\usage{
CLINICAL_GROUPS
}
\description{
Clinical variable groups
}
\keyword{internal}
