% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assign_cancer_type}
\alias{.assign_cancer_type}
\title{Assign cancer type labels (including subtypes) to samples}
\usage{
.assign_cancer_type(sample_ids, requested_cancers)
}
\description{
Assign cancer type labels (including subtypes) to samples
}
\keyword{internal}
