% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.calc_optimal_cutoff}
\alias{.calc_optimal_cutoff}
\title{Calculate optimal cutoff for continuous variable}
\usage{
.calc_optimal_cutoff(data, var_col, time_col, event_col, minprop = 0.1)
}
\description{
Calculate optimal cutoff for continuous variable
}
\keyword{internal}
