% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.categorize_clinical_tcga}
\alias{.categorize_clinical_tcga}
\title{Categorize clinical values}
\usage{
.categorize_clinical_tcga(values, var_name)
}
\description{
Categorize clinical values
}
\keyword{internal}
