% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.create_classic_boxplot}
\alias{.create_classic_boxplot}
\title{Create classic boxplot (Internal)}
\usage{
.create_classic_boxplot(
  data,
  cat_var,
  con_var,
  title,
  xlab,
  ylab,
  stats = NULL
)
}
\description{
Create classic boxplot (Internal)
}
\keyword{internal}
