% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.create_percentage_barplot}
\alias{.create_percentage_barplot}
\title{Create percentage barplot (Internal)}
\usage{
.create_percentage_barplot(
  data,
  var1,
  var2,
  cancer_label,
  stats = NULL,
  var1_label = NULL,
  var2_label = NULL
)
}
\description{
Create percentage barplot (Internal)
}
\keyword{internal}
