% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.custom_gsea_dot}
\alias{.custom_gsea_dot}
\title{Custom GSEA Dot Plot}
\usage{
.custom_gsea_dot(
  res,
  type,
  show.term.num = 15,
  Select.P = "NP",
  cutoff.P = 0.05,
  colors = c("#003c30", "#01665e", "#35978f", "#80cdc1", "#c7eae5", "#f6e8c3", "#dfc27d",
    "#bf812d", "#8c510a", "#543005"),
  y.label.position = "right",
  title = NULL,
  legend.position = "right",
  theme.plot = ggplot2::theme_bw(base_rect_size = 1.5)
)
}
\description{
Custom GSEA Dot Plot
}
\keyword{internal}
