% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scenario.R
\name{.detect_correlation_scenario}
\alias{.detect_correlation_scenario}
\title{Detect Correlation Scenario (Scenarios 1-7)}
\usage{
.detect_correlation_scenario(
  var1_features,
  var2_features,
  var1_types,
  var2_types,
  n_cancers
)
}
\arguments{
\item{var1_features}{Character vector of var1 feature labels}

\item{var2_features}{Character vector of var2 feature labels}

\item{var1_types}{Character vector of var1 types}

\item{var2_types}{Character vector of var2 types}

\item{n_cancers}{Integer. Number of cancers}
}
\value{
List with scenario information
}
\description{
Determines which of 7 correlation scenarios applies based on:
- Number of continuous variables
- Number of categorical variables
- Number of cancers

Scenarios:
1. 1 continuous vs 1 continuous
2. 1 continuous vs multiple continuous (or vice versa)
3. multiple continuous vs multiple continuous
4. 1 categorical vs 1 categorical
5. 1 categorical vs multiple categorical (or vice versa)
6. multiple categorical vs multiple categorical
7. categorical vs continuous (any combination)
}
\keyword{internal}
