% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scenario.R
\name{.detect_enrichment_scenario}
\alias{.detect_enrichment_scenario}
\title{Detect Enrichment Scenario (Scenarios 8-15)}
\usage{
.detect_enrichment_scenario(var_features, var_types, analysis_type)
}
\arguments{
\item{var_features}{Character vector of variable feature labels}

\item{var_types}{Character vector of variable types}

\item{analysis_type}{"genome" or "enrichment"}
}
\value{
List with scenario information
}
\description{
Determines enrichment scenario based on:
- Variable type (continuous or categorical)
- Number of variables
- Analysis type (vs genome or vs enrichment)

Scenarios:
8. 1 categorical vs genome-wide
9. 1 categorical vs enrichment
10. multiple categorical vs genome-wide
11. multiple categorical vs enrichment
12. 1 continuous vs genome-wide
13. 1 continuous vs enrichment
14. multiple continuous vs genome-wide
15. multiple continuous vs enrichment
}
\keyword{internal}
