% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scenario.R
\name{.detect_survival_scenario}
\alias{.detect_survival_scenario}
\title{Detect Survival Scenario (Scenarios 16-17)}
\usage{
.detect_survival_scenario(var_features, var_types, n_cancers)
}
\arguments{
\item{var_features}{Character vector of variable feature labels}

\item{var_types}{Character vector of variable types}

\item{n_cancers}{Integer. Number of cancers}
}
\value{
List with scenario information
}
\description{
Determines survival scenario based on number of variables

Scenarios:
16. 1 variable vs survival (KM + Cox)
17. multiple variables vs survival (Forest plot)
}
\keyword{internal}
