% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.dispatch_correlation_plot}
\alias{.dispatch_correlation_plot}
\title{Dispatch correlation plot based on scenario}
\usage{
.dispatch_correlation_plot(
  scenario_info,
  data,
  stats,
  var1_features,
  var2_features,
  plot_type = "auto"
)
}
\description{
Dispatch correlation plot based on scenario
}
\keyword{internal}
