% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.extract_colname_from_label}
\alias{.extract_colname_from_label}
\title{Extract column name from feature label}
\usage{
.extract_colname_from_label(labels, data)
}
\description{
Converts "THBS2 (Protein, COAD)" → "COAD_THBS2_Protein"
}
\keyword{internal}
