% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filter_subtype_samples}
\alias{.filter_subtype_samples}
\title{Filter samples by subtype}
\usage{
.filter_subtype_samples(cancer_info)
}
\arguments{
\item{cancer_info}{Cancer info from .parse_cancer_type()}
}
\value{
Vector of sample IDs
}
\description{
Filter samples by subtype
}
\keyword{internal}
