% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.generate_filename}
\alias{.generate_filename}
\title{Generate filename for saved plots}
\usage{
.generate_filename(
  analysis,
  var1,
  var1_modal,
  var1_cancers,
  var2 = NULL,
  var2_modal = NULL,
  var2_cancers = NULL
)
}
\description{
Generate filename for saved plots
}
\keyword{internal}
