% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.get_geneset_df}
\alias{.get_geneset_df}
\title{Retrieve gene sets as data.frame from multiple databases}
\usage{
.get_geneset_df(
  type = "GO",
  GO_ont = "BP",
  kegg_category = "pathway",
  msigdb_category = "H",
  hgdisease_source = "do",
  mesh_method = "gendoo",
  mesh_category = "A",
  enrichrdb_library = "Cancer_Cell_Line_Encyclopedia"
)
}
\description{
Retrieve gene sets as data.frame from multiple databases
}
\keyword{internal}
