% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.get_msigdb_geneset}
\alias{.get_msigdb_geneset}
\title{Handle MsigDB category mapping for geneset package}
\usage{
.get_msigdb_geneset(category)
}
\description{
Handle MsigDB category mapping for geneset package
}
\keyword{internal}
