% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_protein_coding_genes}
\alias{.get_protein_coding_genes}
\title{Get protein-coding gene symbols (embedded list)}
\usage{
.get_protein_coding_genes()
}
\description{
Get protein-coding gene symbols (embedded list)
}
\keyword{internal}
