% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.load_clinical_basic}
\alias{.load_clinical_basic}
\title{Load basic clinical data (used by other loaders)}
\usage{
.load_clinical_basic()
}
\value{
Data frame with basic clinical info
}
\description{
Load basic clinical data (used by other loaders)
}
\keyword{internal}
