% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.load_genome_data}
\alias{.load_genome_data}
\title{Load genome-wide data from pre-processed files}
\usage{
.load_genome_data(cancers, genome_modal)
}
\description{
Load genome-wide data from pre-processed files
}
\keyword{internal}
