% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.load_genome_data_tcga}
\alias{.load_genome_data_tcga}
\title{Load genome-wide RNAseq data for enrichment analysis}
\usage{
.load_genome_data_tcga(cancers, rnaseq_type = "log2TPM")
}
\description{
Load genome-wide RNAseq data for enrichment analysis
}
\keyword{internal}
