% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.load_immunecell_modal_tcga}
\alias{.load_immunecell_modal_tcga}
\title{Load immune cell data}
\usage{
.load_immunecell_modal_tcga(vars, cancers, immune_algorithm = NULL)
}
\description{
Load immune cell data
}
\keyword{internal}
