% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.load_methylation_modal_tcga}
\alias{.load_methylation_modal_tcga}
\title{Load methylation data}
\usage{
.load_methylation_modal_tcga(
  vars,
  cancers,
  methylation_region = "Promoter_mean"
)
}
\description{
Load methylation data
}
\keyword{internal}
