% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.load_modal_data}
\alias{.load_modal_data}
\title{Load data for a single modal type}
\usage{
.load_modal_data(vars, modal, cancers, surv_type, var_name)
}
\description{
Load data for a single modal type
}
\keyword{internal}
