% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.load_modal_data_tcga}
\alias{.load_modal_data_tcga}
\title{Load data for a single modal type}
\usage{
.load_modal_data_tcga(
  vars,
  modal,
  cancers,
  surv_type,
  var_name,
  rnaseq_type,
  cnv_type,
  methylation_region,
  immune_algorithm
)
}
\description{
Load data for a single modal type
}
\keyword{internal}
