% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.load_rnaseq_modal_tcga}
\alias{.load_rnaseq_modal_tcga}
\title{Load RNAseq data}
\usage{
.load_rnaseq_modal_tcga(vars, cancers, rnaseq_type = "log2TPM")
}
\description{
Load RNAseq data
}
\keyword{internal}
