% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.map_colname_to_label}
\alias{.map_colname_to_label}
\title{Map column name back to feature label (used in stats.R)}
\usage{
.map_colname_to_label(colnames, colname_vector, label_vector)
}
\description{
Map column name back to feature label (used in stats.R)
}
\keyword{internal}
