% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.map_colname_to_label_simple}
\alias{.map_colname_to_label_simple}
\title{Map column names to feature labels (helper version, not used in stats)}
\usage{
.map_colname_to_label_simple(colnames, modal = NULL)
}
\arguments{
\item{colnames}{Column names}

\item{modal}{Modal type}
}
\value{
Vector of feature labels
}
\description{
Map column names to feature labels (helper version, not used in stats)
}
\keyword{internal}
