% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.match_cancer_parent}
\alias{.match_cancer_parent}
\title{Helper to match samples to requested cancers (main/subtype/combined)}
\usage{
.match_cancer_parent(values, cancer_info)
}
\description{
Helper to match samples to requested cancers (main/subtype/combined)
}
\keyword{internal}
