% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.match_clinical_col_tcga}
\alias{.match_clinical_col_tcga}
\title{Match clinical column name}
\usage{
.match_clinical_col_tcga(search, available_cols)
}
\description{
Match clinical column name
}
\keyword{internal}
