% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.match_immune_cell_col}
\alias{.match_immune_cell_col}
\title{Match immune cell column name}
\usage{
.match_immune_cell_col(search, available, algorithm = NULL)
}
\description{
Match immune cell column name
}
\keyword{internal}
