% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{.match_signature_col_tcga}
\alias{.match_signature_col_tcga}
\title{Match signature column name}
\usage{
.match_signature_col_tcga(search, available_cols)
}
\description{
Match signature column name
}
\keyword{internal}
