% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.merge_list_data}
\alias{.merge_list_data}
\title{Merge list of data frames}
\usage{
.merge_list_data(data_list)
}
\arguments{
\item{data_list}{List of data frames}
}
\value{
Merged data frame
}
\description{
Merge list of data frames
}
\keyword{internal}
