% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.parse_cancer_type}
\alias{.parse_cancer_type}
\title{Parse cancer type input}
\usage{
.parse_cancer_type(cancer_input)
}
\arguments{
\item{cancer_input}{User input cancer type (case insensitive)}
}
\value{
List with type, cancer, subtype info
}
\description{
Parse cancer type input
}
\keyword{internal}
