% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.perform_cox_analysis}
\alias{.perform_cox_analysis}
\title{Perform Cox regression}
\usage{
.perform_cox_analysis(data, var_col, time_col, event_col)
}
\description{
Perform Cox regression
}
\keyword{internal}
