% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.perform_gsea}
\alias{.perform_gsea}
\title{Perform GSEA}
\usage{
.perform_gsea(
  ranked_genes,
  enrich_type = "GO",
  GO_ont = "BP",
  n_workers = 6,
  minSize = 10,
  maxSize = 500,
  kegg_category = "pathway",
  msigdb_category = "H",
  hgdisease_source = "do",
  mesh_method = "gendoo",
  mesh_category = "A",
  enrichrdb_library = "Cancer_Cell_Line_Encyclopedia"
)
}
\description{
Perform GSEA
}
\keyword{internal}
