% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.perform_km_analysis}
\alias{.perform_km_analysis}
\title{Perform Kaplan-Meier analysis}
\usage{
.perform_km_analysis(data, group_col, time_col, event_col)
}
\description{
Perform Kaplan-Meier analysis
}
\keyword{internal}
