% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_categorical_heatmap}
\alias{.plot_categorical_heatmap}
\title{Create categorical heatmap (Internal)}
\usage{
.plot_categorical_heatmap(data, stats, var1_features, var2_features)
}
\description{
Create categorical heatmap (Internal)
}
\keyword{internal}
