% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_categorical_immune_heatmap}
\alias{.plot_categorical_immune_heatmap}
\title{Plot categorical vs immune cells as heatmap with barplot/dotplot
Supports multi-cancer scenarios by composing multiple heatmaps}
\usage{
.plot_categorical_immune_heatmap(data, stats, cat_features, con_features)
}
\description{
Plot categorical vs immune cells as heatmap with barplot/dotplot
Supports multi-cancer scenarios by composing multiple heatmaps
}
\keyword{internal}
