% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_dotplot_paired}
\alias{.plot_dotplot_paired}
\title{DotPlot Paired (Internal)}
\usage{
.plot_dotplot_paired(
  all_stats,
  analysis_type = "DEA",
  cancer_types,
  genome_omics = "Protein",
  is_mutation = FALSE,
  use_mean = FALSE,
  feature_list = NULL,
  method = "pearson",
  top_n = 30
)
}
\description{
DotPlot Paired (Internal)
}
\keyword{internal}
