% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_gsea_matrix}
\alias{.plot_gsea_matrix}
\title{GSEA Matrix DotPlot (Internal)}
\usage{
.plot_gsea_matrix(
  all_gsea_stats,
  enrich_type,
  GO_ont = NULL,
  cancer_types,
  method = NULL,
  use_mean = FALSE,
  feature_list = NULL,
  top_n = 10,
  omics_type = NULL
)
}
\description{
GSEA Matrix DotPlot (Internal)
}
\keyword{internal}
