% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_gsea_paired}
\alias{.plot_gsea_paired}
\title{GSEA Paired DotPlot (Internal)}
\usage{
.plot_gsea_paired(
  gsea_stats,
  var_name,
  omics_type,
  cancer_types,
  enrich_type,
  GO_ont = NULL,
  method = NULL,
  top_n = 15
)
}
\description{
GSEA Paired DotPlot (Internal)
}
\keyword{internal}
