% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_km_cox_combined}
\alias{.plot_km_cox_combined}
\title{KM + Cox Combined Plot (Internal)}
\usage{
.plot_km_cox_combined(
  km_fit,
  cox_model_stats,
  data,
  time_col,
  event_col,
  group_col,
  var_name,
  omics_type,
  cancer_type,
  surv_type,
  var_col = NULL
)
}
\description{
KM + Cox Combined Plot (Internal)
}
\keyword{internal}
