% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_mutation_immune_heatmap}
\alias{.plot_mutation_immune_heatmap}
\title{Plot mutation vs immune cells as heatmap with barplot}
\usage{
.plot_mutation_immune_heatmap(data, stats, cat_features, con_features)
}
\description{
Plot mutation vs immune cells as heatmap with barplot
}
\keyword{internal}
