% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_network}
\alias{.plot_network}
\title{Network Plot (Internal)}
\usage{
.plot_network(
  stats,
  var1_name,
  edge_metric = "correlation",
  query_omics = NULL,
  genome_omics = NULL,
  cancer_type = NULL,
  analysis_type = "correlation",
  method = NULL
)
}
\description{
Network Plot (Internal)
}
\keyword{internal}
