% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_scenario1}
\alias{.plot_scenario1}
\title{Plot for Scenario 1 (Internal)}
\usage{
.plot_scenario1(data, stats, var1_feat, var2_feat, scenario_info)
}
\description{
Plot for Scenario 1 (Internal)
}
\keyword{internal}
