% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_scenario2}
\alias{.plot_scenario2}
\title{Plot for Scenario 2 (Internal)}
\usage{
.plot_scenario2(data, stats, var1_features, var2_features, scenario_info)
}
\description{
Plot for Scenario 2 (Internal)
}
\keyword{internal}
