% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_scenario3}
\alias{.plot_scenario3}
\title{Plot for Scenario 3 (Internal)}
\usage{
.plot_scenario3(data, stats, var1_features, var2_features, scenario_info)
}
\description{
Plot for Scenario 3 (Internal)
}
\keyword{internal}
