% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_scenario4}
\alias{.plot_scenario4}
\title{Plot for Scenario 4 (Internal)}
\usage{
.plot_scenario4(data, stats, var1_features, var2_features, scenario_info)
}
\description{
Plot for Scenario 4 (Internal)
}
\keyword{internal}
