% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_scenario5_6}
\alias{.plot_scenario5_6}
\title{Plot for Scenarios 5-6 (Internal)}
\usage{
.plot_scenario5_6(data, stats, var1_features, var2_features, scenario_info)
}
\description{
Plot for Scenarios 5-6 (Internal)
}
\keyword{internal}
