% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_scenario7}
\alias{.plot_scenario7}
\title{Plot for Scenario 7 (Internal)}
\usage{
.plot_scenario7(data, stats, var1_features, var2_features, scenario_info)
}
\description{
Plot for Scenario 7 (Internal)
}
\keyword{internal}
