% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.run_categorical_enrichment}
\alias{.run_categorical_enrichment}
\title{Run categorical enrichment (DEA-based, Scenarios 8-11)}
\usage{
.run_categorical_enrichment(
  data,
  var_features,
  var1_cancers,
  genome_modal,
  analysis_type,
  enrich_database,
  enrich_ont,
  top_n,
  n_workers,
  kegg_category,
  msigdb_category,
  hgdisease_source,
  mesh_method,
  mesh_category,
  enrichrdb_library
)
}
\description{
Run categorical enrichment (DEA-based, Scenarios 8-11)
}
\keyword{internal}
