% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.run_continuous_enrichment}
\alias{.run_continuous_enrichment}
\title{Run continuous enrichment (Correlation-based, Scenarios 12-15)}
\usage{
.run_continuous_enrichment(
  data,
  var_features,
  var1_cancers,
  genome_modal,
  analysis_type,
  enrich_database,
  enrich_ont,
  method,
  top_n,
  n_workers,
  rnaseq_type,
  kegg_category,
  msigdb_category,
  hgdisease_source,
  mesh_method,
  mesh_category,
  enrichrdb_library
)
}
\description{
Run continuous enrichment (Correlation-based, Scenarios 12-15)
}
\keyword{internal}
