% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.run_survival_forest}
\alias{.run_survival_forest}
\title{Run multiple variables survival analysis (Scenario 17)}
\usage{
.run_survival_forest(
  merged_data,
  var_features,
  var_types,
  time_col,
  event_col,
  surv_type,
  cutoff_type,
  minprop,
  var1_cancers
)
}
\description{
Run multiple variables survival analysis (Scenario 17)
}
\keyword{internal}
