% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.run_survival_single}
\alias{.run_survival_single}
\title{Run single variable survival analysis (Scenario 16)}
\usage{
.run_survival_single(
  merged_data,
  var_feature,
  var_type,
  time_col,
  event_col,
  surv_type,
  cutoff_type,
  minprop,
  palette,
  var_cancers,
  var_col_override = NULL
)
}
\description{
Run single variable survival analysis (Scenario 16)
}
\keyword{internal}
