% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.stats_cor_genome}
\alias{.stats_cor_genome}
\title{Perform genome-wide correlation}
\usage{
.stats_cor_genome(var_data, genome_matrix, var_cancers, method = "pearson")
}
\description{
Perform genome-wide correlation
}
\keyword{internal}
