% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.stats_correlation}
\alias{.stats_correlation}
\title{Calculate Correlation Statistics}
\usage{
.stats_correlation(
  data,
  var1_features,
  var2_features,
  method = "pearson",
  use = "pairwise.complete.obs",
  p_adjust_method = "BH"
)
}
\description{
Calculate Correlation Statistics
}
\keyword{internal}
