% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.stats_dea_genome}
\alias{.stats_dea_genome}
\title{Perform genome-wide DEA using limma}
\usage{
.stats_dea_genome(var_data, genome_matrix, var_cancers, for_enrichment = FALSE)
}
\description{
Perform genome-wide DEA using limma
}
\keyword{internal}
