% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.stats_group_difference}
\alias{.stats_group_difference}
\title{Calculate Group Difference Statistics (Categorical vs Continuous)}
\usage{
.stats_group_difference(
  data,
  cat_features,
  con_features,
  alpha = 0.05,
  p_adjust_method = "BH"
)
}
\description{
Calculate Group Difference Statistics (Categorical vs Continuous)
}
\keyword{internal}
