% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.string_similarity}
\alias{.string_similarity}
\title{Calculate string similarity (Levenshtein distance)}
\usage{
.string_similarity(s1, s2)
}
\arguments{
\item{s1}{String 1}

\item{s2}{String 2}
}
\value{
Similarity score (0-1)
}
\description{
Calculate string similarity (Levenshtein distance)
}
\keyword{internal}
