% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.suggest_similar}
\alias{.suggest_similar}
\title{Suggest similar strings}
\usage{
.suggest_similar(query, candidates, n = 3)
}
\arguments{
\item{query}{Query string}

\item{candidates}{Candidate strings}

\item{n}{Number of suggestions}
}
\value{
Top n similar strings
}
\description{
Suggest similar strings
}
\keyword{internal}
