% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\docType{data}
\name{IMMUNE_CELL_CATEGORIES}
\alias{IMMUNE_CELL_CATEGORIES}
\title{Immune cell categories}
\format{
An object of class \code{list} of length 11.
}
\usage{
IMMUNE_CELL_CATEGORIES
}
\description{
Immune cell categories
}
\keyword{internal}
