% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_cancer_types}
\alias{list_cancer_types}
\title{List Cancer Types}
\usage{
list_cancer_types(show_subtypes = TRUE)
}
\arguments{
\item{show_subtypes}{Logical. Show subtypes (default: TRUE)}
}
\value{
Data frame with cancer information (invisible)
}
\description{
Display all available cancer types (33 main + 32 subtypes = 65 total).
}
\examples{
\dontrun{
list_cancer_types()
list_cancer_types(show_subtypes = FALSE)
}

}
