% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_modalities}
\alias{list_modalities}
\title{List All Available Modalities}
\usage{
list_modalities()
}
\value{
Data frame with modality information (invisible)
}
\description{
Display all omics layers available in SLTCGA with descriptions.
}
\examples{
\dontrun{
list_modalities()
}

}
