% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Login.R
\name{Login}
\alias{Login}
\title{Login in BEST database.}
\usage{
Login(token)
}
\arguments{
\item{token}{A bearer authorization token. You need to log in with a token with the author's permission. E-mail: liuzaoqu@163.com}
}
\description{
Login in BEST database.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
