% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{search_immune_cells}
\alias{search_immune_cells}
\title{Search Immune Cell Types}
\usage{
search_immune_cells(keyword)
}
\arguments{
\item{keyword}{Character. Keyword to search}
}
\value{
Data frame with matching cells (invisible)
}
\description{
Search for immune cell types matching a keyword.
}
\examples{
\dontrun{
search_immune_cells("B_cells")
search_immune_cells("macrophage")
}

}
