% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\docType{data}
\name{SIGNATURE_ALIASES}
\alias{SIGNATURE_ALIASES}
\title{Signature variable aliases}
\format{
An object of class \code{list} of length 37.
}
\usage{
SIGNATURE_ALIASES
}
\description{
Signature variable aliases
}
\keyword{internal}
