% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\docType{data}
\name{SIGNATURE_GROUPS}
\alias{SIGNATURE_GROUPS}
\title{Signature variable groups}
\format{
An object of class \code{list} of length 8.
}
\usage{
SIGNATURE_GROUPS
}
\description{
Signature variable groups
}
\keyword{internal}
