% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pan-Expr-Analysis.R
\name{TCGA.Pan.MG.Expr}
\alias{TCGA.Pan.MG.Expr}
\title{Expression analysis of multiple genes in pancancer.}
\usage{
TCGA.Pan.MG.Expr(
  genes,
  logFC.cutoffs = c(-Inf, -2, -1, -0.5, 0.5, 1, 2, Inf),
  logFC.labels = c("< -2", "-2 - -1", "-1 - -0.5", "-0.5 - 0.5", "0.5 - 1", "1 - 2",
    "> 2"),
  logFC.colors = c("#2166ac", "#4393c3", "#92c5de", "#e5e5e5", "#f4a582", "#d6604d",
    "#b2182b"),
  Select.P = "NP",
  Heatmap.cluster = FALSE,
  ...,
  data.path = ".",
  delete = TRUE
)
}
\arguments{
\item{genes}{A vector with multiple genes.}

\item{logFC.cutoffs}{Cutoffs of logFC.}

\item{logFC.labels}{Labels of logFC cutoffs.}

\item{logFC.colors}{Colors of logFC cutoffs.}

\item{Select.P}{Which P value will be the determinant of significance. NP or FDR.}

\item{Heatmap.cluster}{Whether to cluster the heatmap.}

\item{...}{Additional parameters for Heatmap function.}

\item{data.path}{The path where the downloaded data is stored.}

\item{delete}{Whether to delete the downloaded file.}
}
\description{
Expression analysis of multiple genes in pancancer.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
