% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pan-Mut-Analysis.R
\name{TCGA.Pan.MG.Mut}
\alias{TCGA.Pan.MG.Mut}
\title{Mutation analysis of multiple genes in pancancer.}
\usage{
TCGA.Pan.MG.Mut(
  genes,
  heatmap.colors = c("#FFFFFF", "#fef0d9", "#fdcc8a", "#fc8d59", "#e34a33", "#b30000"),
  ...,
  data.path = ".",
  delete = TRUE
)
}
\arguments{
\item{genes}{A vector with multiple genes.}

\item{heatmap.colors}{Colors for the heatmap.}

\item{...}{Additional parameters for Heatmap function.}

\item{data.path}{The path where the downloaded data is stored.}

\item{delete}{Whether to delete the downloaded file.}
}
\description{
Mutation analysis of multiple genes in pancancer.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
