% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pan-Sur-Analysis.R
\name{TCGA.Pan.MG.Sur}
\alias{TCGA.Pan.MG.Sur}
\title{Survival analysis of multiple genes in pancancer.}
\usage{
TCGA.Pan.MG.Sur(
  genes,
  sur.type = "OS",
  colors = c("#F2F2F2", "#4D76B7", "#D02E20"),
  P.cutoff = 0.05,
  ...,
  data.path = ".",
  delete = TRUE
)
}
\arguments{
\item{genes}{A vector with multiple genes.}

\item{sur.type}{Type of survival infomation. OS, DFS, DSS, and PFS.}

\item{colors}{Colors for Not Significant, Protective and Risky.}

\item{P.cutoff}{Cutoff value of P value.}

\item{...}{Additional parameters for Heatmap function.}

\item{data.path}{The path where the downloaded data is stored.}

\item{delete}{Whether to delete the downloaded file.}
}
\description{
Survival analysis of multiple genes in pancancer.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
