% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pan-CNV-Analysis.R
\name{TCGA.Pan.SG.CNV}
\alias{TCGA.Pan.SG.CNV}
\title{CNV analysis of multiple genes in pancancer.}
\usage{
TCGA.Pan.SG.CNV(
  gene,
  colors = c("#ED6355", "#41A98E", "grey"),
  text.digit = 3,
  text.x.position.just = 0.34,
  data.path = ".",
  delete = TRUE
)
}
\arguments{
\item{gene}{Your gene.}

\item{colors}{Colors for positive, negative and not significant.}

\item{text.digit}{Digits of text.}

\item{text.x.position.just}{X position justification of text.}

\item{data.path}{The path where the downloaded data is stored.}

\item{delete}{Whether to delete the downloaded file.}
}
\description{
CNV analysis of multiple genes in pancancer.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
