% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pan-Expr-Analysis.R
\name{TCGA.Pan.SG.Expr}
\alias{TCGA.Pan.SG.Expr}
\title{Expression analysis of single gene in pancancer.}
\usage{
TCGA.Pan.SG.Expr(
  gene,
  plot.type = "violin",
  colors = c("#41A98E", "#ED6355"),
  add.cancer.without.normalSample = FALSE,
  ...,
  data.path = ".",
  delete = FALSE
)
}
\arguments{
\item{gene}{Your gene.}

\item{plot.type}{Type of the plot; violin or box.}

\item{colors}{colors for normal and tumor groups.}

\item{add.cancer.without.normalSample}{Whether to add cancer types without normal samples.}

\item{...}{Additional parameters for Heatmap function.}

\item{data.path}{The path where the downloaded data is stored.}

\item{delete}{Whether to delete the downloaded file.}
}
\description{
Expression analysis of single gene in pancancer.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
