% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pan-Meth-Analysis.R
\name{TCGA.Pan.SG.Meth}
\alias{TCGA.Pan.SG.Meth}
\title{Methylation analysis of multiple genes in pancancer.}
\usage{
TCGA.Pan.SG.Meth(
  gene,
  type = "promoter",
  colors = c("#7EC09A", "#BB61A1", "grey"),
  text.digit = 3,
  text.x.position.just = 0.34,
  data.path = ".",
  delete = TRUE
)
}
\arguments{
\item{gene}{Your gene.}

\item{type}{Type of methylation sites; global or promoter.}

\item{colors}{Colors for positive, negative and not significant.}

\item{text.digit}{Digits of text.}

\item{text.x.position.just}{X position justification of text.}

\item{data.path}{The path where the downloaded data is stored.}

\item{delete}{Whether to delete the downloaded file.}
}
\description{
Methylation analysis of multiple genes in pancancer.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
