% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pan-Sur-Analysis.R
\name{TCGA.Pan.SG.Sur}
\alias{TCGA.Pan.SG.Sur}
\title{Survival analysis of multiple genes in pancancer.}
\usage{
TCGA.Pan.SG.Sur(
  gene,
  plot.type = "classic",
  sur.type = "OS",
  colors = c("#ED6355", "grey", "#41A98E"),
  rel.width = c(0.5, 1.4, 1.2, 0.6),
  data.path = ".",
  delete = TRUE
)
}
\arguments{
\item{gene}{Your gene.}

\item{plot.type}{Type of the plot; classic or simple.}

\item{sur.type}{Type of survival infomation. OS, DFS, DSS, and PFS.}

\item{colors}{Colors for Risky, Not Significant and Protective.}

\item{rel.width}{Relative width of four columns.}

\item{data.path}{The path where the downloaded data is stored.}

\item{delete}{Whether to delete the downloaded file.}
}
\description{
Survival analysis of multiple genes in pancancer.
}
\author{
Zaoqu Liu; E-mail: liuzaoqu@163.com
}
