% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{TCGA_MAIN_CANCERS}
\alias{TCGA_MAIN_CANCERS}
\title{Main cancer types (33)}
\format{
An object of class \code{character} of length 33.
}
\usage{
TCGA_MAIN_CANCERS
}
\description{
Main cancer types (33)
}
\keyword{internal}
