% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{TCGA_SUBTYPE_MAP}
\alias{TCGA_SUBTYPE_MAP}
\title{Subtype mapping (32 subtypes)}
\format{
An object of class \code{list} of length 27.
}
\usage{
TCGA_SUBTYPE_MAP
}
\description{
Subtype mapping (32 subtypes)
}
\keyword{internal}
