% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_colname_from_label}
\alias{.extract_colname_from_label}
\title{Extract column name from feature label}
\usage{
.extract_colname_from_label(labels, data)
}
\arguments{
\item{labels}{Feature labels in format "GENE (Modal, Cancer)"}

\item{data}{Data frame to search columns in}
}
\value{
Vector of column names
}
\description{
Extract column name from feature label
}
\keyword{internal}
