% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_filename}
\alias{.generate_filename}
\title{Generate filename for plots}
\usage{
.generate_filename(
  analysis,
  var1,
  var1_modal,
  var1_cancers,
  var2 = NULL,
  var2_modal = NULL,
  var2_cancers = NULL
)
}
\arguments{
\item{analysis}{Analysis type (correlation, enrichment, survival)}

\item{var1}{Variable 1}

\item{var1_modal}{Modal 1}

\item{var1_cancers}{Cancers 1}

\item{var2}{Variable 2 (optional)}

\item{var2_modal}{Modal 2 (optional)}

\item{var2_cancers}{Cancers 2 (optional)}
}
\value{
Filename
}
\description{
Generate filename for plots
}
\keyword{internal}
