% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{.stats_association}
\alias{.stats_association}
\title{Calculate Association Statistics (Categorical vs Categorical)}
\usage{
.stats_association(
  data,
  var1_features,
  var2_features,
  alpha = 0.05,
  p_adjust_method = "BH"
)
}
\description{
Calculate Association Statistics (Categorical vs Categorical)
}
\keyword{internal}
